feat: create functinalities for creating the data and show the alert message

+++ database/migrations/2024_06_11_152659_create_campaigns_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('campaigns', function (Blueprint $table) {
             $table->id();
-            $table->integer('category_id');
-            $table->integer('organization_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('organization_id');
             $table->string('title');
             $table->text('description');
             $table->string('image');
